% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_acquisition}
\alias{get_acquisition}
\title{Capitalize Acquisition (-5 years)}
\usage{
get_acquisition(cash_flows, amortized = TRUE)
}
\arguments{
\item{amortized}{\strong{Boolean} Acquisition spending the last 10 years}

\item{cashflows}{\strong{List of numbers} Acquisition spending the 5 years}
}
\value{
\strong{Number} Amortized or Unamortized portion of Acquisition
}
\description{
Capitalize acquisition investments five years back.
}
\examples{
# List of cash flows
# cash_flow <- list(current=123, y1=333, y2=444, y3=444, y4=555, y5=666)
# Calculate unamortized portion of acquisition
# unamortized_aq <- get_acquisition(cash_flows=cash_flow, amortized=FALSE)
# Calculate amortized portion of acquisition
# amortized_aq <- get_acquisition(cash_flows=cash_flow, amortized=FALSE)
}
