% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_amortization_10}
\alias{get_amortization_10}
\title{Capitalize R&D (-10 years)}
\usage{
get_amortization_10(cash_flows, amortized)
}
\arguments{
\item{amortized}{\strong{Boolean} Calculate amortized R&D}

\item{cashflows}{\strong{List of numbers} R&D spending the last 10 years}
}
\value{
\strong{Number} Amortized or Unamortized portion of R&D
}
\description{
Capitalize R&D investments 10 years back (Chemical, Auto/Truck, Drug, Machine, Maritime, Paper, Telecom Equipment, Water Utility, Heavy Manufacturing, Research with patenting).
}
\examples{
# List of cash flows
cash_flow <- list(current=123, y1=333, y2=444, y3=444, y4=555, y5=666, y6=345, y7=556, y8=954, y9=3435, y10=111)
# Calculate unamortized portion of R&D
unamortized_rnd <- get_amortization_10(cash_flow, amortized=FALSE)
# Calculate amortized portion of R&D
amortized_rnd <- get_amortization_10(cash_flow)
}
