% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_beta_start_up}
\alias{get_beta_start_up}
\title{Calculate the first 5Y cost of equity, debt ratio, and cost of capital}
\usage{
get_beta_start_up(
  beta,
  risk_free,
  risk_premium,
  debt,
  shares_out,
  stock_price,
  cost_debt
)
}
\arguments{
\item{beta}{\strong{Number} Approximate firm beta}

\item{risk_free}{\strong{Number} Risk free rate}

\item{risk_premium}{\strong{Number} Risk premium}

\item{debt}{\strong{Number} Firm long and short term debt}

\item{shares_out}{\strong{Number} Number of shares outstanding}

\item{stock_price}{\strong{Number} Current stock price}

\item{cost_debt}{\strong{Number} Synthetic cost of debt (CCC firm)}
}
\value{
\strong{Number} with bottom-up beta
}
\description{
Calculate the first 5Y cost of equity, capital, and debt ratio
}
\examples{
# beta <- get_beta_start_up(risk_free=0.035, risk_premium=0.041, beta=0.77, debt=143, shares_out=81, stock_price=120, cost_debt=0.08)
}
