% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_cash_flow}
\alias{get_cash_flow}
\title{Calculate cash flows}
\usage{
get_cash_flow(after_tax_ebit, reinvestment_rate, time_period, wacc, growth)
}
\arguments{
\item{after_tax_ebit}{A number After-tax EBIT}

\item{reinvestment_rate}{A number Reinvestment rate}

\item{time_period}{A vector of numbers DCF years}

\item{wacc}{A number Cost of Capital}

\item{growth}{A number Estimated growth}
}
\description{
Calculate After-tax EBIT and FCFF flows and NPV (2-stage growth)
}
