% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_cost_capital}
\alias{get_cost_capital}
\title{Calculate Cost of Capital}
\usage{
get_cost_capital(marginal_tax, cost_equity, cost_debt, market_cap, debt)
}
\arguments{
\item{marginal_tax}{\strong{Number} Marginal tax rate}

\item{cost_equity}{\strong{Number} Cost of equity}

\item{cost_debt}{\strong{Number} Cost of debt}

\item{market_cap}{\strong{Number} Firm market capitalization}

\item{debt}{\strong{Number} Firm debt (short and long term)}
}
\value{
\strong{Tibble} Cost of Debt
}
\description{
Calculate Cost of Capital.
}
\examples{
# cost_equity <- get_cost_debt(cost_equity=0.107, cost_debt=0.0929, marginal_tax=0.35, market_cap=500, debt=300)
}
