% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_cost_debt}
\alias{get_cost_debt}
\title{Calculate Cost of Debt}
\usage{
get_cost_debt(risk_free, company_spread, country_spread = 0)
}
\arguments{
\item{risk_free}{A number Risk free rate}

\item{company_spread}{A number Firm company spread}

\item{country_spread}{A number Country spread}
}
\value{
Cost of Debt
}
\description{
Calculate Cost of Debt based on interest coverage rate (EBIT/Interest expense)
}
\examples{
cost_equity <- get_cost_debt(risk_free=0.015, company_spread=0.01)
}
