% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_dynamic_growth}
\alias{get_dynamic_growth}
\title{Calculate Expected Growth rate (changing ROIC)}
\usage{
get_dynamic_growth(
  reinvestment_rate,
  roic_initial,
  roic_target,
  years_target = 5
)
}
\arguments{
\item{reinvestment_rate}{A number Reinvestment rate}

\item{roic_initial}{A number Current year Return on Capital}

\item{roic_target}{A number Target Return on Capital}
}
\description{
Calculate expected growth rate (w/ changing ROIC across years)
}
\examples{
get_stable_growth(reinvestment_rate=0.4, roic_initial=0.2, roic_target=0.6, year_target=5)
}
