% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_fcff}
\alias{get_fcff}
\title{Calculate Free Cash Flow to Firm}
\usage{
get_fcff(after_tax_ebit, depreciation, capex, working_capital)
}
\arguments{
\item{after_tax_ebit}{A number After-Tax EBIT}

\item{depreciation}{A number Depreciation}

\item{capex}{A number Capital Expenditure}

\item{working_capital}{A number Change in Working Capital}
}
\description{
Calculate FCFF (w/ After Tax EBIT)
}
\examples{
get_fcff(after_tax_ebit=2481, depreciation=1914, capex=1659, working_capital=1119)
}
