% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_growth_taxes}
\alias{get_growth_taxes}
\title{Helper function for get_growth_flow}
\usage{
get_growth_taxes(ebit, last_nol, marginal_tax)
}
\arguments{
\item{ebit}{A number EBIT}

\item{last_nol}{A number Last year NOL}

\item{marginal_tax}{A number Marginal tax (industry)}
}
\description{
Calculate Taxes paid
}
