% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_margins}
\alias{get_margins}
\title{Calculate revenues, EBIT, operating margins dynamics (last 5Y)}
\usage{
get_margins(revenues, ebit)
}
\arguments{
\item{revenues}{\strong{List} Revenues the last 5Y}

\item{ebit}{\strong{List} EBIT the last 5Y}
}
\value{
\strong{List} EBIT/Revenue dynamic and margins
}
\description{
Calculate percentage of revenue or EBIT increase/decrease over
the last 5Y. Additionally, calculate dynamics of operating margins (last 5Y)
}
\examples{
# revenues <- list(y4=100,y3=256,y2=546,y1=547, current=850)
# ebit <- list(y4=75,y3=133,y2=400,y1=413, current=750)
# margins <- get_margins(revenues=revenues, ebit=ebit)
}
