% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_net_capex}
\alias{get_net_capex}
\title{Calculate Net CapEx}
\usage{
get_net_capex(
  capex,
  depreciation,
  rnd_expense,
  rnd_amortization,
  acquisition,
  acquisition_amortization = 0
)
}
\arguments{
\item{capex}{A number Current year capital expenditure}

\item{depreciation}{A number Depreciation}

\item{rnd_expense}{A number Current year R&D expense}

\item{rnd_amortization}{A number R&D amortization}

\item{acquisition}{A number Acquisition}
}
\value{
Net CapEx
}
\description{
Calculate and adjust Net Capital Expenditure
}
\examples{
get_net_capex(capex=584, depreciation=484, rnd_expense=1594, rnd_amortization=485, acquisition=2516, ...)
}
