% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_nol}
\alias{get_nol}
\title{Helper function for get_growth_flow}
\usage{
get_nol(nol, ebit)
}
\arguments{
\item{nol}{\strong{Number} NOL}

\item{ebit}{\strong{Number} EBIT}
}
\value{
\strong{Number}
}
\description{
Calculate NOL
}
