% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_operating_assets}
\alias{get_operating_assets}
\title{Calculate Value of Operating Assets}
\usage{
get_operating_assets(fcff, wacc, growth)
}
\arguments{
\item{fcff}{A number Free Cash Flow to Firm}

\item{wacc}{A number Cost of Capital}

\item{growth}{A number Estimate growth}
}
\description{
Calculate Value of Operating Assets
}
\examples{
get_operating_asses(fcff=1617, wacc=12.05, growth=0.05)
}
