% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_revenue_dynamic}
\alias{get_revenue_dynamic}
\title{Get revenue evolution for the past 5Y (start-ups)}
\usage{
get_revenue_dynamic(revenues)
}
\arguments{
\item{revenues}{\strong{List} Revenues the last 5Y}
}
\value{
\strong{List} Percentage list of revenue dynamics
}
\description{
Calculate percentage increase/decrease of start-up revenues.
}
\examples{
# revenues <- list(y4=100,y3=256,y2=546,y1=547, current=850)
# get_revenue_dynamic(revenues = revenues)
}
