% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_roe}
\alias{get_roe}
\title{Calculate ROE}
\usage{
get_roe(
  income,
  rnd_adj = 0,
  equity,
  goodwill,
  goodwill_portion = 0.2,
  rnd_asset = 0,
  eff_tax = 0
)
}
\arguments{
\item{income}{\strong{Number} Current year income}

\item{rnd_adj}{\strong{Number} R&D adjustment}

\item{equity}{\strong{Number} Current year Equity}

\item{goodwill}{\strong{Number} Goodwill}

\item{goodwill_portion}{\strong{Number} Portion of goodwill adjustment}

\item{rnd_asset}{\strong{Number} Unamortized portion of R&D}

\item{eff_tax}{\strong{Number} Effective tax rate}
}
\value{
\strong{Tibble} with ROE and the other parameters
}
\description{
Calculate Return on Equity
}
\examples{
# ROE <- get_roe(income=333, goodwill=5555, equity=4555)
}
