% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_roic}
\alias{get_roic}
\title{Calculate ROIC}
\usage{
get_roic(
  ebit,
  curr_lease_adj = 0,
  rnd_adj = 0,
  eff_tax,
  short_debt,
  long_debt,
  equity,
  goodwill,
  cash,
  last_lease_debt = 0,
  rnd_asset = 0,
  goodwill_portion = 0.2,
  cash_portion = 0.2,
  minority = 0
)
}
\arguments{
\item{ebit}{\strong{Number} Current year EBIT}

\item{curr_lease_adj}{\strong{Number} Lease adjustment to EBIT}

\item{rnd_adj}{\strong{Number} R&D adjustment to EBIT}

\item{eff_tax}{\strong{Number} Effective tax rate (Tax Paid / Taxable Income)}

\item{short_debt}{\strong{Number} Short term debt (Short term interest bearing debt (debt + leases))}

\item{long_debt}{\strong{Number} Long term debt (debt + leases)}

\item{equity}{\strong{Number} Equity}

\item{goodwill}{\strong{Number} Goodwill}

\item{cash}{\strong{Number} Cash}

\item{last_lease_debt}{\strong{Number} Last year capitalized leases}

\item{rnd_asset}{\strong{Number} The unamortized portion of R&D}

\item{goodwill_portion}{\strong{Number} Part of goodwill added to Invested capital}

\item{cash_portion}{\strong{Number} Part of cash added to Invested capital}

\item{minority}{\strong{Number} Minority interest}

\item{current_year_rnd}{\strong{Number} Current year R&D spending}

\item{amortized_rnd}{\strong{Number} Amortized portion of R&D}
}
\value{
\strong{Tibble} with ROIC and other parameters
}
\description{
Calculate Return on Invested Capital.
}
\examples{
# roic <- get_roic(ebit=2033, eff_tax=0.2, short_debt=333,
#                 long_debt=4567, equity=4444, goodwill=333, cash=345)
}
