% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_stable_growth}
\alias{get_stable_growth}
\title{Calculate Expected Growth Rate (stable ROIC)}
\usage{
get_stable_growth(reinvestment_rate, roic)
}
\arguments{
\item{reinvestment_rate}{\strong{Number} Reinvestment rate}

\item{roic}{\strong{Number} Return on Capital}
}
\value{
\strong{Tibble} Expected Growth Rate
}
\description{
Calculate expected growth rate (w/ stable ROIC across years)
}
\examples{
get_stable_growth(reinvestment_rate=0.4, roic=0.2)
}
