% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_stable_operating_assets}
\alias{get_stable_operating_assets}
\title{Calculate FCFF and Value of Operating Assets for stable growth firm}
\usage{
get_stable_operating_assets(
  after_tax_ebit,
  depreciation,
  capex,
  working_capital,
  wacc,
  growth
)
}
\arguments{
\item{after_tax_ebit}{A number After-Tax EBIT (current year)}

\item{depreciation}{A number Depreciation}

\item{capex}{A number Capital Expenditure}

\item{working_capital}{A number Change in Working Capital}

\item{wacc}{A number Cost of Capital}

\item{growth}{A number Growth rate}
}
\description{
Calculate FCFF (w/ After Tax EBIT)
}
\examples{
get_fcff(after_tax_ebit=2481, depreciation=1914, capex=1659, working_capital=1119)
}
