% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_terminal_value}
\alias{get_terminal_value}
\title{Calculate Terminal Value}
\usage{
get_terminal_value(ebit_year_five, stable_growth, wacc, roic)
}
\arguments{
\item{ebit_year_five}{A number After-tax EBIT year 5}

\item{stable_growth}{A number Growth (Risk-free Rate)}

\item{wacc}{A number Cost of Capital}

\item{roic}{A number Return of Capital}
}
\description{
Calculate Value of Terminal Value (2-stage growth)
}
\examples{
get_terminal_value(after_tax_ebit=1300, roic=0.046, wacc=0.1, stable_growth=0.03)
}
