% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_tv}
\alias{get_tv}
\title{Helper function for get_growth_value}
\usage{
get_tv(
  terminal_nol,
  terminal_fcff,
  cost_capital,
  marginal_tax,
  terminal_growth
)
}
\arguments{
\item{terminal_nol}{Terminal NOL}

\item{terminal_fcff}{Terminal FCFF}

\item{cost_capital}{Terminal cost of capital}

\item{marginal_tax}{Marginal tax rate}

\item{terminal_growth}{Terminal growth rate}
}
\value{
Number Terminal value
}
\description{
Calculate terminal value
}
