% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_working_cap}
\alias{get_working_cap}
\title{Calculate working capital}
\usage{
get_working_cap(inventory, accounts_receive, accounts_payable)
}
\arguments{
\item{inventory}{A number Inventory}

\item{accounts_receive}{A number Accounts receivable}

\item{accounts_payable}{A number Accounts payable}
}
\value{
Working capital
}
\description{
Calculate working capital
}
\examples{
get_working_cap(inventory=40, account_receive=80, accounts_payable=40)
}
