% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{rnd_adjustment}
\alias{rnd_adjustment}
\title{Calculate R&D adjustment to EBIT}
\usage{
rnd_adjustment(current_rnd, amortized_rnd)
}
\arguments{
\item{current_rnd}{\strong{Number} Current year R&D expense}

\item{amortized_rnd}{\strong{Number} Amortized R&D}
}
\value{
\strong{Number} R&D adjustment for EBIT
}
\description{
Calculate R&D adjustment to EBIT (for ROIC calculation).
}
\examples{
# See get_amortization function output for amortized R&D
# Current year R&D
# current_rnd <- 1771
# R&D spending last five years
# cash_flow <- list(current=123, y1=333, y2=444, y3=444, y4=555, y5=666)
# Calculate amortized portion of R&D
# amortized_rnd <- get_amortization_5(cash_flow)
# lease_adj <- rnd_adjustment(current_rnd, amortized_rnd)
}
