% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_acquisition}
\alias{get_acquisition}
\title{Capitalize Acquisition (-5 years)}
\usage{
get_acquisition(cash_flows, amortized = TRUE)
}
\arguments{
\item{amortized}{Boolean}

\item{cashflows}{A list of numbers}
}
\value{
Amortized or Unamortized portion of Acquisition
}
\description{
Capitalize acquisition investments
}
\examples{
# Vector of cash flows
cash_flow <- list(current=123, y1=333, y2=444, y3=444, y4=555, y5=666)
# Calculate unamortized portion of Acquisition
unamortized_rnd <- get_acquisition(cash_flow, amortized=FALSE)
# Calculate amortized portion of Acquisition
amortized_rnd <- get_acquisition(cash_flow)
}
