% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_amortization_10}
\alias{get_amortization_10}
\title{Capitalize R&D (-10 years)}
\usage{
get_amortization_10(cash_flows, amortized)
}
\arguments{
\item{amortized}{Boolean}

\item{cashflows}{A list of numbers}
}
\value{
Amortized or Unamortized portion of R&D
}
\description{
Capitalize R&D investments (Chemical, Auto/Truck, Drug, Machine, Maritime, Paper, Telecom Equipment, Water Utility, Heavy Manufacturing, Research with patenting)
}
\examples{
# Vector of cash flows
cash_flow <- list(current=123, y1=333, y2=444, y3=444, y4=555, y5=666)
# Calculate unamortized portion of R&D
unamortized_rnd <- get_amortization_10(cash_flow, amortized=FALSE)
# Calculate amortized portion of R&D
amortized_rnd <- get_amortization_10(cash_flow)
}
