% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_amortization_2}
\alias{get_amortization_2}
\title{Capitalize R&D (-2 years)}
\usage{
get_amortization_2(cash_flows, amortized)
}
\arguments{
\item{amortized}{\strong{Boolean} Calculate amortized R&D}

\item{cashflows}{\strong{List of numbers} R&D spending the last 2 years}
}
\value{
\strong{Number} Amortized or unamortized portion of R&D
}
\description{
Capitalize R&D investments two years back (Non-tech service, Ads, Banks, Restaurant, Retail Store etc.).
}
\examples{
# List of cash flows
cash_flow <- list(current=123, y1=333, y2=444)
# Calculate unamortized portion of R&D
unamortized_rnd <- get_amortization_2(cash_flow, amortized=FALSE)
# Calculate amortized portion of R&D
amortized_rnd <- get_amortization_2(cash_flow)
}
