% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_amortization_3}
\alias{get_amortization_3}
\title{Capitalize R&D (-3 years)}
\usage{
get_amortization_3(cash_flows, amortized)
}
\arguments{
\item{amortized}{Boolean}

\item{cashflows}{A list of numbers}
}
\value{
Amortized or Unamortized portion of R&D
}
\description{
Capitalize R&D investments (Retail tech service, Apparel, Entertainment, Food processing, Household products, Newspapers, Publishing etc.)
}
\examples{
# Vector of cash flows
cash_flow <- list(current=123, y1=333, y2=444, y3=555)
# Calculate unamortized portion of R&D
unamortized_rnd <- get_amortization_3(cash_flow, amortized=FALSE)
# Calculate amortized portion of R&D
amortized_rnd <- get_amortization_3(cash_flow)
}
