% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_beta}
\alias{get_beta}
\title{Calculate Levered Beta / Unlevered beta}
\usage{
get_beta(
  average_beta,
  industry_tax,
  average_de,
  firm_tax,
  firm_debt,
  firm_equity
)
}
\arguments{
\item{average_beta}{A number Industry average beta}

\item{industry_tax}{A number Industry effective tax rate}

\item{average_de}{A number Industry average Debt/Equity}

\item{firm_tax}{A number Firm effective tax}

\item{firm_debt}{A number Firm debt value}

\item{firm_equity}{A number Firm equity value}
}
\value{
Bottom-up beta
}
\description{
Calculate unlevered and levered beta
}
\examples{
beta <- get_beta(average_beta=0.77, industry_tax=0.05, average_de=0.083, ...)
}
