% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_cash_flow}
\alias{get_cash_flow}
\title{Calculate cash flows}
\usage{
get_cash_flow(after_tax_ebit, reinvestment_rate, time_period, wacc, growth)
}
\arguments{
\item{after_tax_ebit}{\strong{Number} After-tax EBIT}

\item{reinvestment_rate}{\strong{Number} Reinvestment rate}

\item{time_period}{\strong{Vector of numbers} DCF years}

\item{wacc}{\strong{Number} Cost of Capital}

\item{growth}{\strong{Number} Estimated growth}
}
\value{
\strong{List of numbers} NPV FCFF, EBIT 5Y, etc.
}
\description{
Calculate After-tax EBIT and FCFF flows and NPV (2-stage growth)
}
\examples{
cash_flows <- get_cash_flow(after_tax_ebit=333, reinvestment_rate=0.3, time_period=c(1,2,3,4,5), wacc=0.08, growth=0.03)
}
