% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_cost_capital}
\alias{get_cost_capital}
\title{Calculate Cost of Capital}
\usage{
get_cost_capital(marginal_tax, cost_equity, cost_debt, equity, debt)
}
\arguments{
\item{marginal_tax}{A number Marginal tax rate}

\item{cost_equity}{A number Cost of equity}

\item{cost_debt}{A number Cost of debt}

\item{equity}{A number Firm equity}

\item{debt}{A number Firm debt}
}
\value{
Cost of Debt
}
\description{
Calculate Cost of Capital
}
\examples{
cost_equity <- get_cost_debt(cost_equity=0.107, cost_debt=0.0929, ...)
}
