% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_cost_equity}
\alias{get_cost_equity}
\title{Calculate Cost of Equity}
\usage{
get_cost_equity(risk_free, beta, risk_premium)
}
\arguments{
\item{risk_free}{A number Risk free rate}

\item{beta}{A number Bottom-up beta}

\item{risk_premium}{A number Risk premium}
}
\value{
Cost of Equity
}
\description{
Calculate Cost of Equity
}
\examples{
cost_equity <- get_cost_equity(risk_free=0.015, beta=0.71, rick_premium=0.041)
}
