% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_dynamic_growth}
\alias{get_dynamic_growth}
\title{Calculate Expected Growth rate (changing ROIC)}
\usage{
get_dynamic_growth(
  reinvestment_rate,
  roic_initial,
  roic_target,
  years_target = 5
)
}
\arguments{
\item{reinvestment_rate}{\strong{Number} Reinvestment rate}

\item{roic_initial}{\strong{Number} Current year Return on Capital}

\item{roic_target}{\strong{Number} Target Return on Capital}
}
\value{
\strong{Tibble} Expected Growth Rate
}
\description{
Calculate expected growth rate (w/ changing ROIC across years), used in case of efficiency claim.
}
\examples{
# get_dynamic_growth(reinvestment_rate=0.4, roic_initial=0.2, roic_target=0.6, year_target=5)
}
