% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_growth_flow}
\alias{get_growth_flow}
\title{Calculate growth firm cash flows}
\usage{
get_growth_flow(
  revenue,
  revenue_growth_trend,
  initial_margin,
  final_margin,
  tax_margin,
  ebit,
  nol
)
}
\arguments{
\item{revenue}{A number Current year revenues}

\item{revenue_growth_trend}{A list of numbers Revenue growth next 10 years}

\item{initial_margin}{A number Current year operating margin}

\item{final_margin}{A number Final year operating margin}

\item{ebit}{A number Current year EBIT}

\item{nol}{A number Net Operating Losses}

\item{marginal_tax}{A number Marginal tax (industry specific)}

\item{depreciation}{A number Depreciation current year}

\item{capex}{A number CapEx current year}

\item{working_capital}{A number Change in working capital current year}
}
\description{
Calculate After-tax EBIT and FCFF flows and NPV (high growth firm)
}
