% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_leases}
\alias{get_leases}
\title{Capitalize future leases}
\usage{
get_leases(lease_flow, beyond, cost_debt)
}
\arguments{
\item{lease_flow}{A list of numbers}

\item{beyond}{A number}

\item{cost_debt}{A number}
}
\value{
List of Debt value of leases AND years embedded in lease (5Y)
}
\description{
Transform future paid leases into debt
}
\examples{
# Paid leases 5Y in the future
cash_flow <- list(y1=333, y2=444, y3=444, y4=555, y5=666)
# Paid leases beyond 5Y
beyond <- 45555
# Indicate the cost of debt (firm rating OR synthetic rating)
cost_debt <- 0.05
# Capitalize future leases
cap_leases <- get_leases(cash_flow, beyond, cost_debt)
# Capitalized leases
cap_leases$debt_value_lease
# Embedded lease years
cap_leases$years_embedded
}
