% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_net_capex}
\alias{get_net_capex}
\title{Calculate Net Capital Expenditures (Net CapEx)}
\usage{
get_net_capex(
  capex,
  depreciation,
  rnd_expense = 0,
  rnd_amortization = 0,
  acquisition = 0,
  lease_last_year = 0,
  lease_current_year = 0,
  acquisition_amortization = 0
)
}
\arguments{
\item{capex}{\strong{Number} Current year capital expenditure}

\item{depreciation}{\strong{Number} Depreciation}

\item{rnd_expense}{\strong{Number} Current year R&D expense}

\item{rnd_amortization}{\strong{Number} R&D amortization}

\item{acquisition}{\strong{Number} Acquisition}

\item{lease_last_year}{\strong{Number} NPV of last year leases}

\item{lease_current_year}{\strong{Number} NPV of current leases}

\item{acquisition_amortization}{\strong{Number} Acquisition amortization}
}
\value{
\strong{Tibble} Net CapEx
}
\description{
Calculate and adjust Net Capital Expenditure.
}
\examples{
get_net_capex(capex=584, depreciation=484)
}
