% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_net_capex}
\alias{get_net_capex}
\title{Calculate Net CapEx}
\usage{
get_net_capex(
  capex,
  depreciation,
  rnd_expense = 0,
  rnd_amortization = 0,
  acquisition = 0,
  lease_last_year = 0,
  lease_current_year = 0,
  acquisition_amortization = 0
)
}
\arguments{
\item{capex}{A number Current year capital expenditure}

\item{depreciation}{A number Depreciation}

\item{rnd_expense}{A number Current year R&D expense}

\item{rnd_amortization}{A number R&D amortization}

\item{acquisition}{A number Acquisition}

\item{lease_last_year}{A number NPV of last year leases}

\item{lease_current_year}{A number NPV of current leases}

\item{acquisition_amortization}{A number Acquisition amortization}
}
\value{
Net CapEx
}
\description{
Calculate and adjust Net Capital Expenditure
}
\examples{
get_net_capex(capex=584, depreciation=484, rnd_expense=1594, rnd_amortization=485, acquisition=2516, ...)
}
