% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_operating_assets}
\alias{get_operating_assets}
\title{Calculate Value of Operating Assets}
\usage{
get_operating_assets(npv_fcff, terminal_value, wacc)
}
\arguments{
\item{npv_fcff}{A number Free Cash Flow to Firm}

\item{terminal_value}{A number Terminal Value}

\item{wacc}{A number Cost of Capital}
}
\description{
Calculate Value of Operating Assets (2-stage growth)
}
\examples{
get_operating_asses(npv_fcff=111, terminal_value=222, wacc=0.065)
}
