% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_roe}
\alias{get_roe}
\title{Calculate ROE}
\usage{
get_roe(income, rnd_adj, equity, goodwill, goodwill_portion = 1, rnd_asset)
}
\arguments{
\item{income}{A number Current year income}

\item{rnd_adj}{A number R&D adjustment}

\item{equity}{A number Current year Equity}

\item{goodwill}{A number Goodwill}

\item{goodwill_portion}{A number Portion of goodwill adjustment (default=0)}

\item{rnd_asset}{A number Unamortized portion of R&D}
}
\value{
A tibble ROE and the other parameters
}
\description{
Calculate Return on Equity
}
\examples{
ROE <- get_roe(income=333, rnd_adj=555, goodwill=5555)
}
