% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_roic}
\alias{get_roic}
\title{Calculate ROIC}
\usage{
get_roic(
  ebit,
  curr_lease_adj = 0,
  rnd_adj = 0,
  eff_tax,
  short_debt,
  long_debt,
  equity,
  goodwill,
  cash,
  last_lease_debt = 0,
  rnd_asset = 0,
  goodwill_portion = 0.2,
  cash_portion = 0.2,
  minority = 0
)
}
\arguments{
\item{ebit}{A number Current year EBIT}

\item{curr_lease_adj}{A number Lease adjustment to EBIT (lease_adjustment function)}

\item{rnd_adj}{A number R&D adjustment to EBIT}

\item{eff_tax}{A number Effective tax rate (Tax Paid / Taxable Income)}

\item{short_debt}{A number Short term debt (Short term interest bearing debt (debt + leases))}

\item{long_debt}{A number Long term debt (Debt + Leases)}

\item{equity}{A number Equity}

\item{goodwill}{A number Goodwill}

\item{cash}{A number Cash}

\item{last_lease_debt}{A number Last year capitalized leases (present value)}

\item{rnd_asset}{A number The unamortized part of R&D}

\item{goodwill_portion}{A number Part of goodwill added to Invested capital (default=0.2)}

\item{cash_portion}{A number Part of cash added to Invested capital (default=0.2)}

\item{minority}{A number Minority interest (default=0)}

\item{current_year_rnd}{A number Current year R&D spending}

\item{amortized_rnd}{A number Amortized portion of R&D}
}
\value{
A tibble ROIC and other parameters
}
\description{
Calculate Return on Invested Capital
}
\examples{
roic <- get_roic(ebit=2033, curr_lease_adj=56763, eff_tax=0.2, short_debt=333,
                 long_debt=4567, equity=4444, ...)
}
