% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_rr}
\alias{get_rr}
\title{Calculate Reinvestment Rate}
\usage{
get_rr(net_capex, working_capital, after_tax_ebit)
}
\arguments{
\item{net_capex}{\strong{Number} Net Capital Expenditure}

\item{working_capital}{\strong{Number} Change in Working Capital}

\item{after_tax_ebit}{\strong{Number} Adjusted EBIT}
}
\value{
\strong{Tibble} Reinvestment rate
}
\description{
Calculate Reinvestment Rate (w/ adjusted EBIT).
}
\examples{
get_rr(net_capex=100, working_capital=100, after_tax_ebit=200)
}
