% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_stable_operating_assets}
\alias{get_stable_operating_assets}
\title{Calculate FCFF and Value of Operating Assets for stable growth firm}
\usage{
get_stable_operating_assets(
  after_tax_ebit,
  net_capex,
  working_capital,
  wacc,
  growth
)
}
\arguments{
\item{after_tax_ebit}{\strong{Number} After-Tax EBIT (current year)}

\item{net_capex}{\strong{Number} Net Capital Expenditures}

\item{working_capital}{\strong{Number} Change in Working Capital}

\item{wacc}{\strong{Number} Cost of Capital}

\item{growth}{\strong{Number} Growth rate}
}
\value{
\strong{Tibble} FCFF
}
\description{
Calculate FCFF (w/ After Tax EBIT)
}
\examples{
get_fcff(after_tax_ebit=2481, depreciation=1914, capex=1659, working_capital=1119, wacc=0.035, growth=0.03)
}
