% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{get_terminal_value}
\alias{get_terminal_value}
\title{Calculate Terminal Value}
\usage{
get_terminal_value(after_tax_ebit, growth, wacc, risk_free)
}
\arguments{
\item{after_tax_ebit}{A number After-tax EBIT}

\item{growth}{A number Estimate growth}

\item{wacc}{A number Cost of Capital}

\item{risk_free}{A number terminal growth rate}

\item{roic}{A number Return of Capital}
}
\description{
Calculate Value of Terminal Value
}
\examples{
get_terminal_value(after_tax_ebit=1300, growth=0.06, wacc=0.1, risk_free=0.03)
}
