% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{lease_adjustment}
\alias{lease_adjustment}
\title{Calculate lease adjustment to EBIT}
\usage{
lease_adjustment(current_lease, debt_value_lease, lease_embedded_years)
}
\arguments{
\item{current_lease}{A number Current lease expense (from IS)}

\item{debt_value_lease}{A number Capitalized leases}

\item{lease_embedded_years}{A number Embedded years in leases (5Y)}
}
\value{
A number of lease adjustment for EBIT
}
\description{
Calculate lease adjustment to EBIT
}
\examples{
See get_leases function output for debt_value_lease and lease_embedded_years
# Current year leases
current_lease <- 555
lease_adj <- lease_adjustment(current_lease, get_leases$debt_value_lease, get_leases$years_embedded)
}
