% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation_funs.R
\name{rnd_adjustment}
\alias{rnd_adjustment}
\title{Calculate R&D adjustment to EBIT}
\usage{
rnd_adjustment(current_rnd, amortized_rnd)
}
\arguments{
\item{current_rnd}{A number Current year R&D expense}

\item{amortized_rnd}{A number Amortized R&D}
}
\value{
A number of R&D adjustment for EBIT
}
\description{
Calculate R&D adjustment to EBIT (for ROIC calculation)
}
\examples{
See get_amortization function output for amortized R&D
# Current year R&D
current_rnd <- 1771
amortized_rnd <- get_amortization_3(...)
lease_adj <- rnd_adjustment(current_rnd, amortized_rnd)
}
